Trival formatting changes.

+++ testsuite/src/test_libblis.c
@@ -137,7 +137,7 @@ void libblis_test_thread_decorator( test_params_t* params, test_ops_t* ops )
 #ifdef BLIS_ENABLE_HPX
 	size_t nt = ( size_t )params->n_app_threads;
 
-	size_t tdata_size = nt *
+	size_t tdata_size = ( size_t )nt *
 	                    ( size_t )sizeof( thread_data_t );
 	thread_data_t* tdata = bli_malloc_user( tdata_size, &r_val );
